deleted and modularized StackNavigator.js
